refactor(toeplitz): remove unused functions
Part of my larger encoding refactor/cleanup.
