+++ encoding/utils/toeplitz/toeplitz.go
@@ -32,49 +32,13 @@ func NewToeplitz(v []fr.Element, fs *fft.FFTSettings) (*Toeplitz, error) {
 	}, nil
 }
 
-// Mutliplication of a matrix and a vector, both elements are Fr Element
-// good info about FFT and Toeplitz,
-// https://alinush.github.io/2020/03/19/multiplying-a-vector-by-a-toeplitz-matrix.html
-func (t *Toeplitz) Multiply(x []fr.Element) ([]fr.Element, error) {
-	cv := t.ExtendCircularVec()
-
-	rv := t.FromColVToRowV(cv)
-	cir := NewCircular(rv, t.Fs)
-
-	xE := make([]fr.Element, len(cv))
-	for i := 0; i < len(x); i++ {
-		xE[i].Set(&x[i])
-	}
-	for i := len(x); i < len(cv); i++ {
-		xE[i].SetZero()
-	}
-
-	product, err := cir.Multiply(xE)
-	if err != nil {
-		return nil, err
-	}
-
-	return product[:len(x)], nil
-}
-
 // Take FFT on Toeplitz vector, coefficient is used for computing hadamard product
 // but carried with multi scalar multiplication
 func (t *Toeplitz) GetFFTCoeff() ([]fr.Element, error) {
-	cv := t.ExtendCircularVec()
-
-	rv := t.FromColVToRowV(cv)
-	cir := NewCircular(rv, t.Fs)
-
-	return cir.GetFFTCoeff()
-}
-
-func (t *Toeplitz) GetCoeff() ([]fr.Element, error) {
-	cv := t.ExtendCircularVec()
-
-	rv := t.FromColVToRowV(cv)
-	cir := NewCircular(rv, t.Fs)
-
-	return cir.GetCoeff()
+	cv := t.extendCircularVec()
+	// TODO(samlaf): why do we convert to row if inside getFFTCoeff we convert back to col?
+	rv := t.fromColVToRowV(cv)
+	return t.getFFTCoeff(rv)
 }
 
 // Expand toeplitz matrix into circular matrix
@@ -90,10 +54,9 @@ func (t *Toeplitz) GetCoeff() ([]fr.Element, error) {
 // [v_4, 0  , v_3, v_2, v_1, v_0, v_6, v_5 ]
 // [v_5, v_4, 0  , v_3, v_2, v_1, v_0, v_6 ]
 // [v_6, v_5, v_4, 0  , v_3, v_2, v_1, v_0 ]
-
-func (t *Toeplitz) ExtendCircularVec() []fr.Element {
+func (t *Toeplitz) extendCircularVec() []fr.Element {
 	E := make([]fr.Element, len(t.V)+1) // extra 1 from extended, equal to 2*dimE
-	numRow := t.GetMatDim()
+	numRow := t.getMatDim()
 	E[0].Set(&t.V[0])
 
 	for i := 1; i < numRow; i++ {
@@ -115,8 +78,7 @@ func (t *Toeplitz) ExtendCircularVec() []fr.Element {
 // if   col Vector is [v_0, v_1, v_2, v_3, 0, v_4, v_5, v_6]
 // then row Vector is [v_0, v_6, v_5, v_4, 0, v_3, v_2, v_1]
 // this operation is involutory. i.e. f(f(v)) = v
-
-func (t *Toeplitz) FromColVToRowV(cv []fr.Element) []fr.Element {
+func (t *Toeplitz) fromColVToRowV(cv []fr.Element) []fr.Element {
 	n := len(cv)
 	rv := make([]fr.Element, n)
 
@@ -130,32 +92,21 @@ func (t *Toeplitz) FromColVToRowV(cv []fr.Element) []fr.Element {
 	return rv
 }
 
-func (t *Toeplitz) GetMatDim() int {
-	return (len(t.V) + 1) / 2
-}
-
-// naive implementation for multiplication. Used for testing
-func (t *Toeplitz) DirectMultiply(x []fr.Element) []fr.Element {
-	numCol := t.GetMatDim()
-
-	n := len(t.V)
-
-	out := make([]fr.Element, numCol)
-	for i := 0; i < numCol; i++ {
-		var sum fr.Element
-		sum.SetZero()
-
-		for j := 0; j < numCol; j++ {
-			idx := (j - i + n) % n
-			var product fr.Element
-			product.Mul(&t.V[idx], &x[j])
-
-			sum.Add(&product, &sum)
-
-		}
-
-		out[i].Set(&sum)
+// Taking FFT on the circular matrix vector
+func (t *Toeplitz) getFFTCoeff(rowV []fr.Element) ([]fr.Element, error) {
+	n := len(rowV)
+	colV := make([]fr.Element, n)
+	for i := 0; i < n; i++ {
+		colV[i] = rowV[(n-i)%n]
 	}
 
-	return out
+	out, err := t.Fs.FFT(colV, false)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (t *Toeplitz) getMatDim() int {
+	return (len(t.V) + 1) / 2
 }
